added full logic in day_logger.h and fixed and added some bugs and new util funcs in other files

+++ src/Club/club_util.cpp
@@ -1,11 +1,11 @@
 #include "club_util.h"
 
 // time_to_int and time_to_str expects and represents the time in the format "hh:mm"
-int time_to_int(const std::string& time) {
-    return std::stoi(time.substr(0, 2)) * 60 + std::stoi(time.substr(3, 2));
+uint32_t time_to_int(const std::string& time) {
+    return static_cast<uint32_t>(std::stoi(time.substr(0, 2)) * 60 + std::stoi(time.substr(3, 2)));
 }
 
-std::string time_to_str(const int time) {
+std::string time_to_str(const uint32_t time) {
     std::string res;
     if (time / 60 < 10) {
         res += "0" + std::to_string(time / 60);
